// src/api/generated/models/recommendationPreferenceResource.ts
/**
 * Generated by orval v7.17.2 🍺
 * Do not edit manually.
 * brew
 * OpenAPI spec version: 0.0.1
 */
import type { RecommendationPreferenceResourceMinRemuneration } from './recommendationPreferenceResourceMinRemuneration';
import type { LanguageResource } from './languageResource';

export interface RecommendationPreferenceResource {
  allow_clt: boolean;
  allow_contractor: boolean;
  allow_internship: boolean;
  allow_on_site: boolean;
  allow_hybrid: boolean;
  allow_remote: boolean;
  on_site_job_radius: number;
  hybrid_jobs_radius: number;
  allow_stack_flexibility: boolean;
  min_remuneration: RecommendationPreferenceResourceMinRemuneration;
  blackListedLanguages: LanguageResource[];
  dev_profile_id: string;
}